Revert codebase to pre-"ObjectTooltip" (that is missing/incomplete) - this should hopefully fix the react frontend building

+++ src/mdx/parsers/w3x/w3i/file.js
@@ -147,7 +147,7 @@ export default class War3MapW3i {
     }
 
     if (this.version >= 27) {
-      this.lua = stream.readInt32();
+      this.lua = stream.readInt8();
     }
 
     for (let i = 0, l = stream.readInt32(); i < l; i++) {
